DB and Site
The DB should be started, it is doing something now, Need data and needs to be able to be queried. Site is starting to take shape and links need to be made to different pages and the pages need to be made. The site needs to be DB merged for queries.